d5p2
I've left a few failed attempts in the code.

Attempt 1: Too slow. Keep a list of rules in memory; filter them by
matching against the line, and spit those into tsort. There are more
rules than input lines, so this was never a good idea.

Attempt 2: Spit the whole set of rules into tsort and extract sublines.
Works on the sample, but the main input cleverly has cycles unless you
use a subset of rules.

Solution: Build an (indirect) hash of rules. The shell doesn't support
subarrays, so we use special variables like rules_NN (like a POSIX
Makefile might). Then, we can extract the rules matching the line with
the ${name:*name} operator (set intersection) once we construct the
right statement (hence eval). We grab these rules for each part of the
line and spit out "$part $follower" where "follower" is the back half of
the rule, but we use a brace-expansion to get there with the ${^name}
flag. If there are no rules (see 13 in my sample), we emit "$part
$part", a dummy rule that means nothing. The whole thing still goes to
tsort.

    hyperfine './part2 input'
    Benchmark 1: ./part2 input
      Time (mean ± σ):      5.861 s ±  0.334 s    [User: 24.798 s, System: 4.315 s]
      Range (min … max):    5.109 s …  6.200 s    10 runs